pyproject.toml
[project]
name = "langfuse-analyst"
version = "0.1.0"
description = "Everything the community Langfuse MCP does, plus analytics. Drop-in replacement with token percentiles, accuracy metrics, failure detection, cost breakdowns, session analytics, latency analysis, and context breach scanning."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "langfuse>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
langfuse-analyst = "langfuse_analyst.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langfuse_analyst"]
